added make_package stage
Summary:
--------
* Added make_package stage
* Added package_build_wrapper.py
* tuscan.py: variable substitutions now performed over entire YAML file
* tuscan.py: implemented command_override
* data containers: changed representation in YAML file

Details:
--------
This commit introduces the make_package stage. This stage is not called
directly from tuscan.py like the other stages. Rather, it is called
(once for each package) from the ninja file output by the deps_to_ninja
stage. The actual invocation of the make_package stage is wrapped up in
the package_build_wrapper.py script.

This stage required several changes to the data representation:

* Several aspects of YAML files are now parametric over the toolchain.
  Thus, tuscan.py now globally substitutes all instances of
  ${TOOLCHAIN} with the toolchain name, rather than in just one key.

* Since the make_package stage is not invoked from tuscan.py, the
  command_override key was introduced to facilitate invoking that stage.

* The data containers YAML file has been refactored to move even more
  data away from tuscan.py, and for better consistency. In particular,
  if a data container specifies its switch name as "foo", then
  containers will always be passed that data container's name as
  --foo-volume and the data container's mountpoint as --foo-directory.
  This is necessary in this commit to ensure that the make_package stage
  uses consistent naming with respect to all the other stages.

+++ tuscan.py
@@ -22,8 +22,37 @@ from os import getcwd, listdir
 from os.path import dirname, isfile, join
 from re import sub
 from subprocess import call
+from string import Template
 from sys import stdout, stderr
 from yaml import load
+
+
+def substitute_vars(data_structure, args):
+    """Substitute variables in YAML data files with values.
+
+    Data stored in YAML files can contain variables which need to be
+    resolved at runtime. This method returns the dict or list passed in
+    as the data_structure argument, with all substitutions applied.
+    """
+    if isinstance(data_structure, basestring):
+        ret = Template(data_structure)
+        ret = Template(ret.safe_substitute(TOOLCHAIN=args.toolchain))
+        ret = Template(ret.safe_substitute(TOUCH_DIR=args.touch_dir))
+
+        # Up to this point, ret is a Template. Turn it into a string
+        # with no-op substitute
+        ret = ret.safe_substitute()
+    elif isinstance(data_structure, list):
+        ret = []
+        for e in data_structure:
+            ret.append(substitute_vars(e, args))
+    elif isinstance(data_structure, dict):
+        ret = {}
+        for k, v in data_structure.items():
+            ret[k] = substitute_vars(v, args)
+    else: raise RuntimeError("Impossible type with value " +
+                             str(data_structure))
+    return ret
 
 
 def run_ninja(args, ninja_file):
@@ -47,7 +76,7 @@ def create_build_file(args, ninja):
     stages.write_build_recipies()
 
     # The top-level build rule
-    ninja.build("build", "phony", touch("run", "custom_repository",
+    ninja.build("build", "phony", touch("run", "make_package",
                                         args))
 
 
@@ -73,10 +102,7 @@ class DataContainers(object):
                          " 'data_containers.yaml'.\n")
             exit(1)
 
-        self.containers = containers
-        for cont in containers:
-            cont["name"] = sub("__TOOLCHAIN__", args.toolchain,
-                               cont["name"])
+        self.containers = substitute_vars(containers, self.args)
         self.data_container_sanity_checks()
 
 
@@ -151,7 +177,7 @@ class Stages(object):
                              " contain a deps.yaml file.\n")
                 exit(1)
 
-        self.stages = stages
+        self.stages = substitute_vars(stages, self.args)
         self.normalise_containers()
         self.container_sanity_checks()
 
@@ -177,33 +203,42 @@ class Stages(object):
 
             commands = []
 
-            docker = "docker run --rm "
+            if "command_override" in sta["run"]:
+                main_command = sta["run"]["command_override"]
+            else:
+                main_command = "docker run --rm "
 
-            for cont in data_containers_needed_by(sta,
-                            self.data_containers):
-                docker += "-v " + cont["mountpoint"]
-                docker += " --volumes-from " + cont["name"] + " "
+                for cont in data_containers_needed_by(sta,
+                                self.data_containers):
+                    main_command += "-v " + cont["mountpoint"]
+                    main_command += " --volumes-from " + cont["name"]
+                    main_command += " "
 
-            docker += " -t " + sta["name"]
-            docker += " --output-directory " + self.args.touch_dir
-            docker += " --toolchain " + self.args.toolchain
+                main_command += " -t " + sta["name"]
+                main_command += " --output-directory "
+                main_command += self.args.touch_dir
+                main_command += " --toolchain " + self.args.toolchain
 
-            for cont in data_containers_needed_by(sta,
-                            self.data_containers):
-                if "switch" in cont:
-                    docker += (" " + cont["switch"] +
-                               " " + cont["mountpoint"])
+                for cont in data_containers_needed_by(sta,
+                                self.data_containers):
+                    if "switch" in cont:
+                        main_command += (" --" + cont["switch"] +
+                                         "-directory "
+                                         + cont["mountpoint"])
+
+                        main_command += (" --" + cont["switch"] +
+                                         "-volume "
+                                         + cont["name"])
 
             if sta["run"]["stdout"]:
-                docker += " >" + sta["run"]["stdout"]
+                main_command += " >" + sta["run"]["stdout"]
             if sta["run"]["stderr"]:
-                docker += " 2>" + sta["run"]["stderr"]
+                main_command += " 2>" + sta["run"]["stderr"]
 
-            commands.append(docker)
+            commands.append(main_command)
 
             if "post_exit" in sta["run"]:
                 cmd = sta["run"]["post_exit"]
-                cmd = sub("__TOUCH_DIR__", self.args.touch_dir, cmd)
                 cmd = sub("\$", "$$", cmd)
                 commands.append(cmd.strip())
 
@@ -291,12 +326,6 @@ class Stages(object):
 
             if not "data_containers" in sta["run"]["dependencies"]:
                 sta["run"]["dependencies"]["data_containers"] = []
-
-            tmp = []
-            for cont in sta["run"]["dependencies"]["data_containers"]:
-                tmp.append(sub("__TOOLCHAIN__", self.args.toolchain,
-                           cont))
-            sta["run"]["dependencies"]["data_containers"] = tmp
 
             if not "stages" in sta["run"]["dependencies"]:
                 sta["run"]["dependencies"]["stages"] = []
